Added GetActiveOrders and GetCompletedOrders Api calls. Updated order id data type for Cancel order request Added JsonSerializerOptions to parse name as case insensitive as cancel order response parsing was breaking.

+++ BitmartSpot/Models/TransactionOrder/CancelOrder/Request.cs
@@ -2,10 +2,10 @@
 namespace BitmartApiClient.Models.TransactionOrder.CancelOrder {
     public class Request {
         [JsonProperty("order_id")]
-        public int OrderId { get; set; }
-        public static Request New(int anOrderId) => new Request(anOrderId);
+        public long OrderId { get; set; }
+        public static Request New(long anOrderId) => new Request(anOrderId);
 
-        private Request(int anOrderId) {
+        private Request(long anOrderId) {
             OrderId = anOrderId;
         }
     }
